+++ lib/integrations/eslint.js
@@ -57,11 +57,7 @@ EslintIntegration._checkFiles = function(files, callback) {
   console.log('Checking linting...');
   if (!files.length) {
     console.log('No changes files since provided date. All ok.');
-    return callback(undefined, {
-      errors: 0,
-      warnings: 0,
-      formatted: ''
-    });
+    return callback();
   }
 
   console.log('Files: ', files);
